got the item list to send to client from server. now about to do item tab do display information

+++ src/Client/Item_Info.java
@@ -1,10 +1,15 @@
 package Client;
 
+import java.util.ArrayList;
+
 public class Item_Info {
     String item_name; // read in from input file
-    int item_min_bid_price; // read in from input file and gets replaced by the client bids
+    double item_min_bid_price; // read in from input file and gets replaced by the client bids
     String client_winner;
-    int client_win_price;
+    double client_win_price;
     boolean available_or_sold; // t if sold, false if available
-    int buy_now_price;
+    double buy_now_price;
+
+
 }
+
